First attempt to create new skill 'achieve'.
The goal is to create a skill that allow the player to kill a dying monster with a hit which cannot be missed. This skill has the internal number 102.
A new animation has been added to the character and this skill can be use with 'Space' key. This skill will be modified further and the new animation can be removed if something better is found.